separate file for getting  generic button rows data
